requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
requests>=2.26.0
beautifulsoup4>=4.10.0
feedparser>=6.0.8
python-dotenv>=0.19.0
google-api-python-client
freegpt
bs4
chromadb
sentence-transformers==2.2.2
faiss-cpu
openai==0.28
torch>=2.0.0
transformers==4.31.0
sentencepiece==0.1.99
accelerate==0.20.3
aiohttp
asyncio